Add exercise 1
I learned more about the useReducer API. When you say React.useReducer
you get an array with two elements: the first being the current object
(state) and the second being a function that takes in two parameters.
The first of these params is the current value for the object (state)
and the second is _something_ that can be used to change that object -
it could just be the new value or it could be a function. This function
returns a value that the original object (state) is then set to.

It took me a bit to process how the dispatch param and the reducer
function work together. The dispatch param calls the reducer function
and passes along the current value of state to it as the first param.
It felt a bit strange that I was only passing 1 param along when the
function takes 2 params - but it's cause dispatch handles that for us.