Add exercise 2
I learned about the existence of useCallback and useMemo in React - and
that they can be used to increase performance by making it so React
understands when certain pieces change vs. when they don't.

I also learned that you have to be careful about making API calls when
components may unmount as they can result in an error where they're
trying to dispatch or do something to something that no longer exists.
There are a variety of ways of addressing this - but it seems like the
most common way is to make a mounted ref and set it to true at the
beginning of a useEffect and set it to false in the return value (which
is called when the component is unmounted).