Add exercise 3
I learned that there is a React concept called "Context" that lets you
save state without having to directly pass it down the chain.

I learned that you should probably generally avoid that in favor of
other things - such as just passing things down or utilizing the
children property to sort of hide away this idea.

I learned /re-learned that you need to be careful when extracting
components out into other components or functions as you have to make
sure to pass along the children or things won't work well!