Add exercise 4
I learned that there is another type of effect in React called
useLayoutEffect. This occurs _before_ the browser paints the screen and
is synchronous. That synchronous part is important because it means that
if you make a useLayoutEffect take a super long time - the browser will
literally just not display any updates until that's done - even if
plenty of stuff is good to go.

This can have a benefit, though - as if you want to make certain changes
before a user sees them - you can utilize this effect. For instance,
let's say you had a chat box like there is on Twitch and you wanted the
user to always see the latest message - if you added the message and
then used an effect to scroll down - the user would see there's
another message before the scroll happened and then it would snap down
after a bit to the latest spot - which is a strange user experience.