Add exercise 5
I learned about the existence of the useImperativeHandle hook and have a
vague idea of how to use it. I learned that you can't really make a ref
for a functional component - but you can sort of get around that by
wrapping it in a forwardRef call - where you can then call
useImperativeHandle in to update the current value of the ref.