Add exercise 6
I learned that you can make custom hooks more easy to understand in the
React dev tools via the use of React's useDebugValue hook. It will allow
us to provide labels for items that would all just look the same
without any value. Worth noting is that this is _only_ for custom hooks
- you can't use this in a functional component or anything like that.